+++ src/com/intellij/idea/plugin/hybris/common/HybrisConstants.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,7 +21,6 @@ import com.intellij.facet.FacetTypeId
 import com.intellij.idea.plugin.hybris.facet.YFacet
 import com.intellij.idea.plugin.hybris.facet.YFacetType
 import com.intellij.openapi.util.io.FileUtilRt
-import java.io.File
 
 object HybrisConstants {
 
@@ -146,6 +145,7 @@ object HybrisConstants {
     const val PLATFORM_TOMCAT_6_DIRECTORY = "tomcat-6"
     const val PLATFORM_TOMCAT_DIRECTORY = "tomcat"
     const val PLATFORM_LIBRARY_GROUP = "Platform Bootstrap"
+    const val PLATFORM_DATABASE_DRIVER_LIBRARY = "Database Drivers"
 
     const val NODE_MODULES_DIRECTORY = "node_modules"
     const val JS_TARGET_DIRECTORY = "jsTarget"