+++ src/com/intellij/idea/plugin/hybris/project/configurators/JavaCompilerConfigurator.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -16,17 +17,21 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.intellij.idea.plugin.hybris.project.configurators;
+package com.intellij.idea.plugin.hybris.project.configurators
 
-import com.intellij.openapi.module.ModifiableModuleModel;
-import com.intellij.openapi.project.Project;
-import org.jetbrains.annotations.NotNull;
+import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.project.Project
 
-public interface SearchScopeConfigurator {
+interface JavaCompilerConfigurator {
 
-    void configure(
-        @NotNull final Project project,
-        @NotNull final ModifiableModuleModel model
-    );
+    fun configure(
+        descriptor: HybrisProjectDescriptor,
+        project: Project,
+        cache: HybrisConfiguratorCache
+    )
 
+    companion object {
+        val instance: JavaCompilerConfigurator = ApplicationManager.getApplication().getService(JavaCompilerConfigurator::class.java)
+    }
 }