+++ src/com/intellij/idea/plugin/hybris/project/configurators/RunConfigurationConfigurator.kt
@@ -1,6 +1,7 @@
 /*
- * This file is part of "hybris integration" plugin for Intellij IDEA.
+ * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -20,14 +21,19 @@ package com.intellij.idea.plugin.hybris.project.configurators
 import com.intellij.idea.plugin.hybris.project.configurators.impl.DebugRunConfigurationConfigurator
 import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.progress.ProgressIndicator
 import com.intellij.openapi.project.Project
 
 interface RunConfigurationConfigurator {
 
-    fun configure(hybrisProjectDescriptor: HybrisProjectDescriptor, project: Project, cache: HybrisConfiguratorCache)
+    fun configure(
+        indicator: ProgressIndicator,
+        hybrisProjectDescriptor: HybrisProjectDescriptor,
+        project: Project,
+        cache: HybrisConfiguratorCache
+    )
 
     companion object {
-        val debugInstance: RunConfigurationConfigurator = ApplicationManager.getApplication().getService(DebugRunConfigurationConfigurator::class.java)
-        val testInstance: RunConfigurationConfigurator? = ApplicationManager.getApplication().getService(TestRunConfigurationConfigurator::class.java)
+        val instance: RunConfigurationConfigurator = ApplicationManager.getApplication().getService(DebugRunConfigurationConfigurator::class.java)
     }
 }