+++ src/com/intellij/idea/plugin/hybris/project/configurators/SpringConfigurator.kt
@@ -1,7 +1,7 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
  * Copyright (C) 2014-2016 Alexander Bartash <AlexanderBartash@gmail.com>
- * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,23 +18,31 @@
  */
 package com.intellij.idea.plugin.hybris.project.configurators
 
-import com.intellij.idea.plugin.hybris.project.configurators.impl.ReadOnlyContentRootConfigurator
-import com.intellij.idea.plugin.hybris.project.configurators.impl.RegularContentRootConfigurator
+import com.intellij.idea.plugin.hybris.project.configurators.impl.DefaultSpringConfigurator
+import com.intellij.idea.plugin.hybris.project.configurators.impl.DummySpringConfigurator
+import com.intellij.idea.plugin.hybris.project.descriptors.HybrisProjectDescriptor
 import com.intellij.idea.plugin.hybris.project.descriptors.ModuleDescriptor
 import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.externalSystem.service.project.IdeModifiableModelsProvider
 import com.intellij.openapi.progress.ProgressIndicator
-import com.intellij.openapi.roots.ModifiableRootModel
 
-interface ContentRootConfigurator {
+interface SpringConfigurator {
+
+    fun process(
+        indicator: ProgressIndicator,
+        hybrisProjectDescriptor: HybrisProjectDescriptor,
+        moduleDescriptors: Map<String, ModuleDescriptor>
+    )
 
     fun configure(
         indicator: ProgressIndicator,
-        modifiableRootModel: ModifiableRootModel,
-        moduleDescriptor: ModuleDescriptor
+        hybrisProjectDescriptor: HybrisProjectDescriptor,
+        moduleDescriptors: Map<String, ModuleDescriptor>,
+        modifiableModelsProvider: IdeModifiableModelsProvider
     )
 
     companion object {
-        val instance: ContentRootConfigurator = ApplicationManager.getApplication().getService(RegularContentRootConfigurator::class.java)
-        val readOnlyInstance: ContentRootConfigurator = ApplicationManager.getApplication().getService(ReadOnlyContentRootConfigurator::class.java)
+        val instance: DefaultSpringConfigurator? = ApplicationManager.getApplication().getService(DefaultSpringConfigurator::class.java)
+        val dummyInstance: DummySpringConfigurator = ApplicationManager.getApplication().getService(DummySpringConfigurator::class.java)
     }
 }