+++ src/com/intellij/idea/plugin/hybris/project/configurators/TestRunConfigurationConfigurator.java
@@ -1,4 +1,0 @@
-package com.intellij.idea.plugin.hybris.project.configurators;
-
-public interface TestRunConfigurationConfigurator extends RunConfigurationConfigurator {
-}